+++ addon/export.py
@@ -56,18 +56,45 @@ class DSC_ParkingSpaceObject(xodr.Object):
 
 from scenariogeneration.xosc.position import ClothoidSpline, ClothoidSplineSegment
 
+def register_lane_type_walking():
+    """
+        The OpenDRIVE lane type "walking" (which replaces the deprecated
+        "sidewalk") is not part of the LaneType enumeration of the
+        scenariogeneration package (as of version 0.16.5). Register it
+        dynamically so that it can be exported with the correct type name.
+    """
+    if hasattr(xodr.LaneType, 'walking'):
+        return xodr.LaneType.walking
+    try:
+        member = object.__new__(xodr.LaneType)
+        member._name_ = 'walking'
+        member._value_ = max(m.value for m in xodr.LaneType) + 1
+        member.__objclass__ = xodr.LaneType
+        # Needs to be inserted into the class dict before the member map since
+        # the enum metaclass refuses to overwrite existing members
+        type.__setattr__(xodr.LaneType, 'walking', member)
+        xodr.LaneType._member_map_['walking'] = member
+        xodr.LaneType._value2member_map_[member._value_] = member
+        xodr.LaneType._member_names_.append('walking')
+        return member
+    except Exception as e:
+        print('Could not register OpenDRIVE lane type "walking", '
+              'falling back to "sidewalk":', e)
+        return xodr.LaneType.sidewalk
+
 mapping_lane_type = {
     'driving': xodr.LaneType.driving,
     #'bidirectional': xodr.LaneType.bidirectional,
     #'bus': xodr.LaneType.bus,
     'stop': xodr.LaneType.stop,
     'parking': xodr.LaneType.parking,
-    #'biking': xodr.LaneType.biking,
+    'biking': xodr.LaneType.biking,
     #'restricted': xodr.LaneType.restricted,
     #'roadWorks': xodr.LaneType.roadWorks,
     'border': xodr.LaneType.border,
-    # TODO (missing) 'curb': xodr.LaneType.curb,
+    'curb': xodr.LaneType.curb,
     #'sidewalk': xodr.LaneType.sidewalk,
+    'walking': register_lane_type_walking(),
     'shoulder': xodr.LaneType.shoulder,
     'median': xodr.LaneType.median,
     'entry': xodr.LaneType.entry,
@@ -1001,6 +1028,7 @@ class DSC_OT_export(bpy.types.Operator):
                                       color=mapping_road_mark_color[color],
                                       marking_weight=mapping_road_mark_weight[weight])
             if width > 0.0:
+                width = helpers.round_float_property(width)
                 # Use the <type><line> element to define the road mark width
                 if marking_type == 'broken':
                     line = xodr.RoadLine(width=width, length=line_length,
@@ -1012,10 +1040,34 @@ class DSC_OT_export(bpy.types.Operator):
                 road_mark.add_specific_road_line(line)
             return road_mark
 
+    def add_lane_height(self, lane, lane_type, height_inner, height_curb):
+        '''
+            Add the height record lifting a lane above the road surface and
+            return the height of its outer edge. A curb lane lifts all lanes
+            behind it, all other lanes keep the height they are lifted to.
+
+            OpenDRIVE interpolates linearly between the inner and the outer
+            height of a lane, hence a curb with different inner and outer
+            heights would describe a diagonal face. To get the vertical face of
+            a square curb (form B) the whole curb lane is put on the raised
+            level instead, which turns the lane border towards the road into a
+            height step, just like in the generated mesh.
+        '''
+        if lane_type == 'curb':
+            height_outer = height_inner + height_curb
+            height_inner = height_outer
+        else:
+            height_outer = height_inner
+        if height_inner != 0.0 or height_outer != 0.0:
+            lane.add_height(helpers.round_float_property(height_inner),
+                            helpers.round_float_property(height_outer))
+        return height_outer
+
     def create_lanes(self, obj):
         lanes = xodr.Lanes()
         road_mark_line_length = obj.get('road_mark_line_length', 3.0)
         road_mark_line_space = obj.get('road_mark_line_space', 6.0)
+        height_curb = helpers.round_float_property(obj.get('height_curb', 0.0))
         road_mark = self.get_road_mark(obj['lane_center_road_mark_type'],
                                        obj['lane_center_road_mark_weight'],
                                        obj['lane_center_road_mark_color'],
@@ -1025,6 +1077,9 @@ class DSC_OT_export(bpy.types.Operator):
         lane_center = xodr.standard_lane(rm=road_mark)
         lane_center.add_roadmark
         lanesection = xodr.LaneSection(0,lane_center)
+        # Lane heights accumulate from the road center outwards
+        height_left = 0.0
+        height_right = 0.0
         for idx in range(obj['lanes_left_num']):
             a,b,c,d = self.get_lane_width_coefficients(obj['lanes_left_widths_start'][idx],
                 obj['lanes_left_widths_end'][idx], obj['geometry_total_length'])
@@ -1037,6 +1092,8 @@ class DSC_OT_export(bpy.types.Operator):
                                            road_mark_line_length,
                                            road_mark_line_space)
             lane.add_roadmark(road_mark)
+            height_left = self.add_lane_height(lane, obj['lanes_left_types'][idx],
+                                               height_left, height_curb)
             lanesection.add_left_lane(lane)
         for idx in range(obj['lanes_right_num']):
             a,b,c,d = self.get_lane_width_coefficients(obj['lanes_right_widths_start'][idx],
@@ -1050,13 +1107,15 @@ class DSC_OT_export(bpy.types.Operator):
                                            road_mark_line_length,
                                            road_mark_line_space)
             lane.add_roadmark(road_mark)
+            height_right = self.add_lane_height(lane, obj['lanes_right_types'][idx],
+                                                height_right, height_curb)
             lanesection.add_right_lane(lane)
         lanes.add_lanesection(lanesection)
         lanes.add_laneoffset(xodr.LaneOffset(0,
-                                             obj['lane_offset_coefficients']['a'],
-                                             obj['lane_offset_coefficients']['b'] / obj['geometry_total_length'],
-                                             obj['lane_offset_coefficients']['c'] / obj['geometry_total_length']**2,
-                                             obj['lane_offset_coefficients']['d'] / obj['geometry_total_length']**3))
+            helpers.round_float_property(obj['lane_offset_coefficients']['a']),
+            helpers.round_float_property(obj['lane_offset_coefficients']['b'] / obj['geometry_total_length']),
+            helpers.round_float_property(obj['lane_offset_coefficients']['c'] / obj['geometry_total_length']**2),
+            helpers.round_float_property(obj['lane_offset_coefficients']['d'] / obj['geometry_total_length']**3)))
 
         return lanes
 
@@ -1074,7 +1133,8 @@ class DSC_OT_export(bpy.types.Operator):
             b = 0.0
             c = 3.0 / length_road**2 * (width_end - width_start)
             d = -2.0 / length_road**3 * (width_end - width_start)
-        return a, b, c, d
+        return helpers.round_float_property(a), b, \
+            helpers.round_float_property(c), helpers.round_float_property(d)
 
     def link_lanes(self, roads):
         '''